Added shader for highlighting objects
Added a shader that highlights objects with a glowing outline to indicate that the object is interactable.

Added a script that slowly rotates objects.

Made improvements to the door script. (now requires you press a button before entry, also the door gets highlighted if you approach with a key.)

Added placeholder UI elements for dialogue boxes.