Added string to paddleball
Disabled the "BallSpawn" script so that the paddleball no longer instantiates a new ball every time it is pulled out.

Items pulled toward the player (like boxes) no longer get auto deleted (Except keys).

Added string that connects the paddle to the ball. This is achieved with two scripts, one is for MainCamera (to render the string) and the other script goes on the string gameobject.